requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyTelegramBotAPI==3.8.1
requests==2.25.1
urllib3==1.26.6
Werkzeug==2.0.1
pymongo~=3.12.0
dnspython==2.1.0
python-telegram-bot==13.7
python-dotenv==0.19.0
Pillow~=8.3.1